Extracted fetching clusters logic and added API error handling
